requirements_gcp.txt
# Google cloud packages
google-cloud-bigquery
db-dtypes
google-cloud-storage

# database
sqlalchemy
#psycopg2
psycopg2-binary

# data science
pandas
fastparquet

# databricks
databricks-cli

# vscode
black
pylint